+++ CHANGELOG.md
@@ -12,6 +12,8 @@ v5.2.0 - TBD
 * Adds `Texture#upload()` and `Texture#download()` to allow direct manipulation
   of a texture's RGBA bitmap.
 * Adds `Thread#on_shutDown()` for running code when a thread terminates.
+* Adds `SSj.now()` and `SSj.lostTime()` which allow very accurate timing when
+  running under SpheRun.
 * Adds support for render-to-texture by allowing `Surface` objects to be used
   as textures, without calling `.toTexture()` first.
 * Adds `[Symbol.toStringTag]` and `.constructor` for Sphere v2 objects.